Migrated src/components/EventStats/* from Jest to Vitest (#2997)
* Refactor: Migrated src/components/EventStats/* from Jest to Vitest

* Refactor: Removed duplicate mock implementation in EventStatsWrapper.spec.tsx
